refactor: extract `heightController` hooks from `alldayGridRow` and `otherGridRow`
